+++ fesm2020/ondewo-nlu-client-angular.mjs
@@ -25725,10 +25725,10 @@ class SessionFilter {
                     _instance.minEntityTypesConfidenceMax = _reader.readFloat();
                     break;
                 case 8:
-                    _instance.earliest = _reader.readFloat();
+                    _instance.earliest = _reader.readDouble();
                     break;
                 case 9:
-                    _instance.latest = _reader.readFloat();
+                    _instance.latest = _reader.readDouble();
                     break;
                 case 10:
                     _instance.minNumberTurns = _reader.readInt32();
@@ -25889,10 +25889,10 @@ class SessionFilter {
             _writer.writeFloat(7, _instance.minEntityTypesConfidenceMax);
         }
         if (_instance.earliest) {
-            _writer.writeFloat(8, _instance.earliest);
+            _writer.writeDouble(8, _instance.earliest);
         }
         if (_instance.latest) {
-            _writer.writeFloat(9, _instance.latest);
+            _writer.writeDouble(9, _instance.latest);
         }
         if (_instance.minNumberTurns) {
             _writer.writeInt32(10, _instance.minNumberTurns);
@@ -26573,10 +26573,10 @@ class SessionInfo {
                     _instance.minEntityTypesConfidence = _reader.readFloat();
                     break;
                 case 6:
-                    _instance.earliest = _reader.readFloat();
+                    _instance.earliest = _reader.readDouble();
                     break;
                 case 7:
-                    _instance.latest = _reader.readFloat();
+                    _instance.latest = _reader.readDouble();
                     break;
                 case 8:
                     _instance.numberTurns = _reader.readInt32();
@@ -26688,10 +26688,10 @@ class SessionInfo {
             _writer.writeFloat(5, _instance.minEntityTypesConfidence);
         }
         if (_instance.earliest) {
-            _writer.writeFloat(6, _instance.earliest);
+            _writer.writeDouble(6, _instance.earliest);
         }
         if (_instance.latest) {
-            _writer.writeFloat(7, _instance.latest);
+            _writer.writeDouble(7, _instance.latest);
         }
         if (_instance.numberTurns) {
             _writer.writeInt32(8, _instance.numberTurns);
